Run Heimdall in its own ECS cluster
There are no services in this integration environment that need this
fix, but this is how Heimdall will be deployed elsewhere, so it's better
to have it the same here.

Also made it so that you can specify separate subnets for the internal
and external load balancers for the Heimdall module.